Preserve tag padding in rewritten QuickTime files
Canon CR3s add padding to each of the tags in the QuickTime container. When
ExifTool reads and rewrites the file, the padding is discarded (because the
length of data within the tag is shorter.) Pad tags with nul bytes back to its
original size when the -QuickTimePreservePadding flag is set, which keeps the
majority of the original file contents intact.